Added the function to iterate over the cubes and safe the triangles in scene
I put it in scene as the main class is just a "mailman", the grid class is a helping class to build the initial grid of cubes, its gonna be handled in the scene class though as we handle the triangles there aswell, i deleted unnecessary methods form the grid class, and commentated some function so that we still have them if i made a mistake. Up till now there seems to be no error. Unless this gitignore  file which is bothering me.